+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openpolicy/core
 
+## 0.0.12
+
+### Patch Changes
+
+- Unified Config
+
 ## 0.0.11
 
 ### Patch Changes